fixed issues that Google found, related to the client being offline in the middle of a Firebase document fetch
