+++ app/src/main/java/com/mikewarren/speakify/utils/log/IResultLoggable.kt
@@ -1,0 +1,16 @@
+package com.mikewarren.speakify.utils.log
+
+import android.util.Log
+
+interface IResultLoggable: ITaggable {
+    fun getSuccessLogMessage() : String
+    fun getFailureLogMessage() : String
+
+    fun logSuccessResult() {
+        Log.d(TAG, getSuccessLogMessage())
+    }
+
+    fun logFailureResult(e: Exception) {
+        Log.e(TAG, getFailureLogMessage(), e)
+    }
+}
